+++ .gitignore
@@ -1,3 +1,4 @@
 cgroup-warden
 .env
 dist/
+vagrant/
