--- cgroup-warden.service
@@ -1,27 +1,0 @@
-[Unit]
-Description=cgroup-warden
-Documentation=https://github.com/chpc-uofu/cgroup-warden
-
-Wants=sssd.service network.target network-online.target
-After=sssd.service network.target network-online.target
-
-[Service]
-
-Type=simple
-Restart=always
-RestartSec=60
-
-CPUAccounting=true
-MemoryAccounting=true
-
-# TODO: update to proper environment file
-EnvironmentFile=/opt/cgroup-warden.conf
-
-# TODO: update to proper working directory
-WorkingDirectory=-/opt/cgroup-warden/
-
-# TODO: update to proper binary path
-ExecStart=/opt/cgroup-warden/cgroup-warden
-
-[Install]
-WantedBy=multi-user.target
